+++ lib/ch/result.ex
@@ -3,18 +3,19 @@ defmodule Ch.Result do
   Result struct returned from any successful query. Its fields are:
 
     * `command` - An atom of the query command, for example: `:select`, `:insert`
-    * `rows` - The result set. One of:
-      - a list of lists, each inner list corresponding to a
-        row, each element in the inner list corresponds to a column
-      - raw iodata when the response is not automatically decoded, e.g. `x-clickhouse-format: CSV`
     * `num_rows` - The number of fetched or affected rows
+    * `types` - The response types
+    * `rows` - A list of lists, each inner list corresponding to a row, each element in the inner list corresponds to a column
+    * `data` - The raw iodata from the response
   """
 
-  defstruct [:command, :num_rows, :rows, :headers]
+  defstruct [:command, :num_rows, :types, :rows, :data]
 
   @type t :: %__MODULE__{
           command: Ch.Query.command() | nil,
           num_rows: non_neg_integer | nil,
-          rows: [[term]] | iodata | nil
+          types: [term] | nil,
+          rows: [[term]] | nil,
+          data: iodata
         }
 end